backend/requirements.txt
asgiref==3.8.1
Django==5.2.2
django-environ==0.12.0
djangorestframework==3.16.0
psycopg2-binary==2.9.10
setuptools==78.1.1
sqlparse==0.5.3
wheel==0.45.1
